Cleanup collection of prometheus metrics
